+++ src/game/tests.rs
@@ -409,6 +409,7 @@ mod laces;
 mod land_and_ability_layers;
 mod land_characteristics;
 mod landwalk;
+mod mana_ability_costs;
 mod mana_and_costs;
 mod mana_vault;
 mod modal_effects;